server/requirements/base.txt
django
djangorestframework
djangorestframework-jwt
django-cors-headers
django_extensions
mysqlclient
coreapi